fix(edit_view): redraw upon font changes
This makes changes to the font instantaneous (once xi sends them to
us) instead of just changing the font once we redraw for other
reasons (e.g. when resizing the window). This had been a little
irritating at times, e.g. when changing the font in the PrefsWin,
but the EditView is only redrawn when changing back to the MainWin.