fix(edit_view): don't set upper limits during draw/scroll but during update
Having both handle_da_draw and scroll_to mess with the horizontal bar doesn't
work so nicely, so do it in update instead. We also pay attention to selections
now, which is nice when using 'find' to search for a word (previously search results
could be right at the edge of the view, or even a bit out of the view, which
was rather confusing). It'd be nice if we could stop calculating the max_width in
get_text_size and Xi told us the width, but for now we'll have to live with
this.

See https://github.com/xi-editor/xi-editor/issues/479