refactor(gxi): spawn xrl's Client in connect_activate
It's heavily recommend by GTK upstream to do as little as possible in main()
and instead move the setup into connect_activate. That's because when we already
have an instace of gxi open and want to open a new tab (e.g. by doing `gxi new-file`)
GTK will execute main() and then connect_open(). Before this change we'd spawn
xi-core only to kill it once connect_open has been run.