feat(editview): add UI to set tab_size,auto_indent and insert_spaces per EditView
We recently introduced the ability to set per-syntax configuration of tab_size and insert_spaces.
This brought a problem with it: We couldn't just keep the previous buttons in the hamburger menu
of the MainWin, since it'd be very confusing if we had a instance-wide button for these settings,
but also allowed syntax specific settings (which may differ from EditView to EditView). As such we
need to have per-EditView settings, which default to what the syntax config says (if there is one),
otherwise default to the instance-wide default but also allow per-view config.

Also move the context_menu to its own glade file. Since glade doesn't know how to deal with GMenus
it just removes them.