fix(tau): don't block on getting view_id in connect_open, fixes opening multiple files
Instead we now spawn a thread which is then blocked instead of our mainthread to
get the view_id. We don't run this on the tokio event loop since this may stall it,
which we want to avoid. Also move the shutdown logic for the tokio Runtime into
connect_shutdown instead of doing it in main_win.win.connect_delete_event, since
shutdown is guarenteed to be called no matter what (e.g. also in app.quit()).