fix(tau): don't start drawing the background at negative X's
Previously we would just subtract the right margin with the current
value of the horizontal scrollbar if the right margin is enabled. This
works fine unless hadj's value is bigger than that of right margin,
in which case we'd start drawing at negative values (but still only
drawed for the EditView's width, so the background wasn't wide enough).

fixes #391