fix(editview|tau): properly send back results of measure_width
Previously we appended all results of measure_width to an array and then
sent xi-core an array of that array (so the "shell" array always was of
the size 1). However, if Xi wants us to measure the width of multiple
styles, it expects us to send back an array which contains one array
of f32s (line widths) per stylle.
Also actually pay attention to what style xi-editor has set for a string
when measuring width, as that could impact the width.

fixes #423